OLH-2619: Restrict actions workflows to repo read permissions
Set the GITHUB_TOKEN permissions to read only for the workflows which run
the tests and check for pinned action versions.

If a workflow doesn't have any permissions defined it inherits from the
repo permissions. Since this repo was created before Feb 2023, the
default permissions are read/write.

These workflows don't need any write permissions so we can restrict them
to the minimum for safety.